requirements.txt
faster-whisper>=1.0.0
sounddevice>=0.4.6
numpy>=1.24.0
pynput>=1.7.6
pystray>=0.19.4
Pillow>=10.0.0
pyperclip>=1.8.2
psutil>=5.9.0
customtkinter>=5.2.0
requests>=2.31.0
